requirements.txt
Appium-Python-Client
# psycopg2
